tests/requirements_base.txt
# Base testing requirements
google-adk>=1.7.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
python-dotenv>=1.0.0
pydantic>=2.0.0
aiofiles>=23.0.0
